+++ src/npctalk.cpp
@@ -441,9 +441,9 @@ void npc::talk_to_u()
  if (d.topic_stack.back() == TALK_NONE) {
   d.topic_stack.back() = pick_talk_topic(&(g->u));
  }
- 
+
  moves -= 100;
- 
+
  if(g->u.has_disease("deaf")) {
   g->add_msg(_("%s tries to talk to you, but you're deaf!"), name.c_str());
   if(d.topic_stack.back() == TALK_MUG) {
@@ -452,7 +452,7 @@ void npc::talk_to_u()
   }
   return;
  }
- 
+
  decide_needs();
 
  d.win = newwin(FULL_SCREEN_HEIGHT, FULL_SCREEN_WIDTH,
@@ -2222,7 +2222,7 @@ Tab key to switch lists, letters to pick items, Enter to finalize, Esc to quit\n
   getting_yours[i] = false;
  }
 
- int  cash = cost;// How much cash you get in the deal (negative = losing money)
+ long cash = cost;// How much cash you get in the deal (negative = losing money)
  bool focus_them = true; // Is the focus on them?
  bool update = true;  // Re-draw the screen?
  int  them_off = 0, you_off = 0;// Offset from the start of the list
